+++ huffman.h
@@ -1,4 +1,5 @@
 #include <stdbool.h>
+#include <stdint.h>
 
 struct huffman_node {
 	int weight;
@@ -7,7 +8,7 @@ struct huffman_node {
 	struct huffman_node *right;
 };
 
-struct huffman_node *huffman_build_tree(void **values, int count);
+struct huffman_node *huffman_build_tree(void **values, uint64_t count);
 
 void *huffman_lookup (struct huffman_node *tree, unsigned char *bits,
 		      int *bits_read, bool print);
